+++ src/runtime/resolver.rs
@@ -778,7 +778,7 @@ impl TryFromRant for SelectorMode {
     }
   }
 
-  fn is_rant_optional() -> bool {
+  fn is_optional_param_type() -> bool {
     false
   }
 }