+++ src/stdlib/boolean.rs
@@ -5,7 +5,7 @@ use super::*;
 /// Returns the logical AND of the operands.
 pub(crate) fn and(vm: &mut VM, (lhs, rhs, extra): (bool, bool, VarArgs<bool>)) -> RantStdResult {
   let result = (lhs && rhs) && extra.iter().all(|b| *b);
-  vm.cur_frame_mut().write_value(RantValue::Boolean(result));
+  vm.cur_frame_mut().write(result);
   Ok(())
 }
 
@@ -14,7 +14,7 @@ pub(crate) fn and(vm: &mut VM, (lhs, rhs, extra): (bool, bool, VarArgs<bool>)) -
 /// Returns the logical OR of the operands.
 pub(crate) fn or(vm: &mut VM, (lhs, rhs, extra): (bool, bool, VarArgs<bool>)) -> RantStdResult {
   let result = (lhs || rhs) || extra.iter().any(|b| *b);
-  vm.cur_frame_mut().write_value(RantValue::Boolean(result));
+  vm.cur_frame_mut().write(result);
   Ok(())
 }
 
@@ -22,7 +22,7 @@ pub(crate) fn or(vm: &mut VM, (lhs, rhs, extra): (bool, bool, VarArgs<bool>)) ->
 ///
 /// Gets the inverse of the operand.
 pub(crate) fn not(vm: &mut VM, val: bool) -> RantStdResult {
-  vm.cur_frame_mut().write_value(RantValue::Boolean(!val));
+  vm.cur_frame_mut().write(!val);
   Ok(())
 }
 
@@ -30,6 +30,6 @@ pub(crate) fn not(vm: &mut VM, val: bool) -> RantStdResult {
 ///
 /// Retirms the logical XOR of the operands.
 pub(crate) fn xor(vm: &mut VM, (lhs, rhs): (bool, bool)) -> RantStdResult {
-  vm.cur_frame_mut().write_value(RantValue::Boolean(lhs ^ rhs));
+  vm.cur_frame_mut().write(lhs ^ rhs);
   Ok(())
 }